// src/app/interfaces/measuresDetails.ts
export interface MandatoryMeasures {
    mandatoryMeasureid: number;
    mandatoryMeasureName: string;
  }

export interface MandatoryMeasureOptions{
    mandatoryMeasureId: number;
    mandatoryMeasureName: string;
    mandatoryDetailsId: number;
    mandatoryMeasureDetails: string;
    mandatoryMeasureOptionId: number;
    mandatoryMeasureOptionName: string;
}